requirements.txt
django>=3.2
paho-mqtt
